# src/main/resources/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Grunner
  license:
    name: MIT
#servers:
#  - url: http://grunner.io/v1
paths:
  /batches:
    get:
      summary: List user's batch & latest status
      operationId: listBatchs
      parameters:
        - name: status
          in: query
          description: Only list batches of specified status
          required: false
          schema:
            $ref: '#/components/schemas/BatchStatus'
      responses:
        200:
          description: The list of (matching) batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BatchStatusUpdate"
    post:
      summary: Submit a new batch
      operationId: enqueueBatch
      requestBody:
        description: The script to execute
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        202:
          description: Unique identifier of new batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchId"
        400:
          description: The script was not syntactically valid
  /batches/{batchId}:
    delete:
      summary: Cancel a batch
      operationId: cancelBatch
      parameters:
        - name: batchId
          in: path
          required: true
          description: Id of batch to cancel
          schema:
            $ref: '#/components/schemas/BatchId'
      responses:
        204:
          description: Batch was cancelled
        403:
          description: Batch was not 'Pending' and could not cancelled
        404:
          description: Specified batch could not be found within user scope

components:
  schemas:
    BatchId:
      description: Unique identifier of a batch
      type: string
    BatchStatus:
      type: string
      description: Possible states of a batch
      enum:
        - Pending
        - Executing
        - Completed
        - Cancelled
        - Failed
    BatchInfo:
      description: All info of a batch
      properties:
        id:
          $ref: '#/components/schemas/BatchId'
        script:
          description: The script to execute
          type: string
        owner:
          description: Username of the submitter
          type: string
    BatchStatusUpdate:
      properties:
        batch-id:
          $ref: '#/components/schemas/BatchId'
        status:
          $ref: '#/components/schemas/BatchStatus'
        timestamp:
          description: The time at which the batch was submitted
          type: string
          format: date-time
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic

security:
  - BasicAuth:
      - read
      - write